Clear the pre-existing detekt and spotless failures
./gradlew check failed on this branch before any of today's changes:
detekt MaxLineLength in GoalTelemetryStoreTest, LongMethod on
ParallelCodeReviewRunner.parentPrompt, TooManyFunctions in
AgentRunAdapters, and two spotless violations.

Split parentPrompt's mode instruction and register contract into their
own builders, moved the Cursor stream decoding out of AgentRunAdapters
into CursorAgentRunDecoding, and applied spotless. No behavior change.
